# Transparently speed up step_bs