--- a/R/bs.R
+++ b/R/bs.R
@@ -113,7 +113,8 @@ splines_wrapper <- function(x, args, type = c("bs", "ns")) {
   # the original implementations issue a warning.
   if (!is.null(args$df) && is.null(args$knots) && args$df - degree - intercept >= 1L) {
     num_knots <- args$df - degree - intercept
-    knots <- unname(quantile(x, seq_len(num_knots) / (num_knots + 1L), na.rm = TRUE))
+    ok <- !is.na(x) & x > boundary[1L] & x < boundary[2L]
+    knots <- unname(quantile(x[ok], seq_len(num_knots) / (num_knots + 1L)))
   } else {
     knots <- numeric()
   }