--- a/R/bs.R
+++ b/R/bs.R
@@ -100,11 +100,9 @@ step_bs_new <-
     )
   }
 
-splines_wrapper <- function(x, args, type = c("bs", "ns")) {
-  type <- match.arg(type)
-
+bs_wrapper <- function(x, args) {
   # Only do the parameter computations from splines::bs() / splines::ns(), don't evaluate at x.
-  degree <- as.integer(args$degree %||% switch(type, bs = 3L, ns = 1L))
+  degree <- as.integer(args$degree %||% 3L)
   intercept <- as.logical(args$intercept %||% FALSE)
   # This behaves differently from splines::ns() if length(x) is 1
   boundary <- sort(args$Boundary.knots) %||% range(x)
@@ -121,22 +119,19 @@ splines_wrapper <- function(x, args, type = c("bs", "ns")) {
 
   # Only construct the data necessary for splines_predict
   out <- matrix(NA, ncol = degree + length(knots) + intercept, nrow = 1L)
-  class(out) <- c(type, "basis", "matrix")
+  class(out) <- c("bs", "basis", "matrix")
   attr(out, "knots") <- knots
   attr(out, "Boundary.knots") <- boundary
   attr(out, "intercept") <- intercept
-  if (type == "bs") {
-    attr(out, "degree") <- degree
-  }
+  attr(out, "degree") <- degree
   out
 }
 
-splines_predict <- function(object, x) {
+bs_predict <- function(object, x) {
   xu <- unique(x)
   ru <- predict(object, xu)
   res <- ru[match(x, xu), ]
-  copy_attrs <- c("class", if (inherits(object, "bs")) "degree" else NULL,
-                  "knots", "Boundary.knots", "intercept")
+  copy_attrs <- c("class", "degree", "knots", "Boundary.knots", "intercept")
   attributes(res)[copy_attrs] <- attributes(ru)[copy_attrs]
   res
 }
@@ -149,7 +144,7 @@ prep.step_bs <- function(x, training, info = NULL, ...) {
   opt <- x$options
   opt$df <- x$deg_free
   opt$degree <- x$degree
-  obj <- lapply(training[, col_names], splines_wrapper, opt, type = "bs")
+  obj <- lapply(training[, col_names], bs_wrapper, opt)
   for (i in seq(along.with = col_names))
     attr(obj[[i]], "var") <- col_names[i]
   step_bs_new(
@@ -177,7 +172,7 @@ bake.step_bs <- function(object, new_data, ...) {
     cols <- (strt):(strt + new_cols[i] - 1)
     orig_var <- attr(object$objects[[i]], "var")
     bs_values[, cols] <-
-      splines_predict(object$objects[[i]], getElement(new_data, i))
+      bs_predict(object$objects[[i]], getElement(new_data, i))
     new_names <-
       paste(orig_var, "bs", names0(new_cols[i], ""), sep = "_")
     colnames(bs_values)[cols] <- new_names