--- a/R/bs.R
+++ b/R/bs.R
@@ -100,7 +100,7 @@ step_bs_new <-
     )
   }
 
-bs_wrapper <- function(x, args) {
+bs_statistics <- function(x, args) {
   # Only do the parameter computations from splines::bs() / splines::ns(), don't evaluate at x.
   degree <- as.integer(args$degree %||% 3L)
   intercept <- as.logical(args$intercept %||% FALSE)
@@ -144,7 +144,7 @@ prep.step_bs <- function(x, training, info = NULL, ...) {
   opt <- x$options
   opt$df <- x$deg_free
   opt$degree <- x$degree
-  obj <- lapply(training[, col_names], bs_wrapper, opt)
+  obj <- lapply(training[, col_names], bs_statistics, opt)
   for (i in seq(along.with = col_names))
     attr(obj[[i]], "var") <- col_names[i]
   step_bs_new(