--- a/tests/testthat/test_ns.R
+++ b/tests/testthat/test_ns.R
@@ -31,6 +31,23 @@ test_that('correct basis functions', {
   carbon_ns_te_exp <- predict(carbon_ns_tr_exp, biomass_te$carbon)
   hydrogen_ns_te_exp <- predict(hydrogen_ns_tr_exp, biomass_te$hydrogen)
 
+  expect_equal(
+    unname(attr(carbon_ns_tr_exp, "knots")),
+    attr(with_ns$steps[[1]]$objects$carbon, "knots")
+  )
+  expect_equal(
+    unname(attr(carbon_ns_tr_exp, "Boundary.knots")),
+    attr(with_ns$steps[[1]]$objects$carbon, "Boundary.knots")
+  )
+  expect_equal(
+    unname(attr(hydrogen_ns_tr_exp, "knots")),
+    attr(with_ns$steps[[1]]$objects$hydrogen, "knots")
+  )
+  expect_equal(
+    unname(attr(hydrogen_ns_tr_exp, "Boundary.knots")),
+    attr(with_ns$steps[[1]]$objects$hydrogen, "Boundary.knots")
+  )
+
   carbon_ns_tr_res <- as.matrix(with_ns_pred_tr[, grep("carbon", names(with_ns_pred_tr))])
   colnames(carbon_ns_tr_res) <- NULL
   hydrogen_ns_tr_res <- as.matrix(with_ns_pred_tr[, grep("hydrogen", names(with_ns_pred_tr))])
@@ -50,23 +67,6 @@ test_that('correct basis functions', {
   dimnames(carbon_ns_te_res) <- NULL
   dimnames(hydrogen_ns_tr_res) <- NULL
   dimnames(hydrogen_ns_te_res) <- NULL
-
-  expect_equal(
-    unname(attr(carbon_ns_tr_exp, "knots")),
-    attr(with_ns$steps[[1]]$objects$carbon, "knots")
-  )
-  expect_equal(
-    unname(attr(carbon_ns_tr_exp, "Boundary.knots")),
-    attr(with_ns$steps[[1]]$objects$carbon, "Boundary.knots")
-  )
-  expect_equal(
-    unname(attr(hydrogen_ns_tr_exp, "knots")),
-    attr(with_ns$steps[[1]]$objects$hydrogen, "knots")
-  )
-  expect_equal(
-    unname(attr(hydrogen_ns_tr_exp, "Boundary.knots")),
-    attr(with_ns$steps[[1]]$objects$hydrogen, "Boundary.knots")
-  )
 
   expect_equal(carbon_ns_tr_res, carbon_ns_tr_exp)
   expect_equal(carbon_ns_te_res, carbon_ns_te_exp)