fix #574

Testing script for benchmark:

```
devtools::load_all()
library(tibble)
set.seed(123)
v_u <- rnorm(100)
x <- tibble(
  v = sample(v_u, 1000000, TRUE)
)
bench::mark(new_code = recipe(x) %>%
  step_bs(v) %>%
  prep()
)
```

Current master on R 4.0.3, Windows 10 x64, i9-10900K:

```
# A tibble: 1 x 13
  expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result   memory                   time         gc              
  <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>   <list>                   <list>       <list>          
1 old_code      573ms    573ms      1.75     637MB     10.5     1     6      573ms <recipe> <Rprofmem[,3] [693 x 3]> <bch:tm [1]> <tibble [1 x 3]>
```

PR code on same machine:

```
# A tibble: 1 x 13
  expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result   memory                     time         gc              
  <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>   <list>                     <list>       <list>          
1 new_code     71.5ms   71.5ms      14.0     119MB     83.9     1     6     71.5ms <recipe> <Rprofmem[,3] [1,411 x 3]> <bch:tm [5]> <tibble [5 x 3]>
```