@AshesITR 2021-01-30
To check that the behaviour didn't change, I checked the objects from both runs using `waldo::compare()`.

```
> waldo::compare(old_result, new_result)
`attr(old$steps[[1]]$terms[[1]], '.Environment')` is <env:00000236F22898F0>
`attr(new$steps[[1]]$terms[[1]], '.Environment')` is <env:00000236FB83C4F8>

`old$steps[[1]]$id`: "bs_yls7H"
`new$steps[[1]]$id`: "bs_Fz6IO"
```