@AshesITR 2021-01-30
@juliasilge I can't find the Rout.fail logs to investigate. Can you help me by pointing me where to look?
Thanks!