@AshesITR 2021-01-30
cc @juliasilge @topepo 

If you like the implementation, I can include an analogous version for `step_ns` - the knot "prep" code in `splines::ns()` is the same as in `splines::ns()`