@AshesITR 2021-01-30
Benchmark results for ns:

```
devtools::load_all()
library(tibble)
set.seed(123)
v_u <- rnorm(100)
x <- tibble(
  v = sample(v_u, 1000000, TRUE)
)
bench::mark(old_code = recipe(x) %>%
  step_ns(v) %>%
  prep()
)
```

```
# A tibble: 1 x 13
  expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result   memory                   time         gc              
  <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>   <list>                   <list>       <list>          
1 new_code      105ms    119ms      6.98     142MB     14.0     4     8      573ms <recipe> <Rprofmem[,3] [876 x 3]> <bch:tm [4]> <tibble [4 x 3]>
```

```
# A tibble: 1 x 13
  expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result   memory                   time         gc              
  <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>   <list>                   <list>       <list>          
1 old_code      657ms    657ms      1.52     977MB     7.61     1     5      657ms <recipe> <Rprofmem[,3] [901 x 3]> <bch:tm [1]> <tibble [1 x 3]>
```