@topepo 2021-02-01
Why do we need a function that replicates the `splines` code? Can't we just pre-compute and use the functions as-is? 

For example: 

``` r
library(tidymodels)
#> ── Attaching packages ────────────────────────────────────── tidymodels 0.1.2 ──
#> ✓ broom     0.7.3           ✓ recipes   0.1.15.9000
#> ✓ dials     0.0.9.9000      ✓ rsample   0.0.8.9000 
#> ✓ dplyr     1.0.3           ✓ tibble    3.0.6      
#> ✓ ggplot2   3.3.3           ✓ tidyr     1.1.2      
#> ✓ infer     0.5.3           ✓ tune      0.1.2.9000 
#> ✓ modeldata 0.1.0.9000      ✓ workflows 0.2.1      
#> ✓ parsnip   0.1.4.9000      ✓ yardstick 0.0.7.9000 
#> ✓ purrr     0.3.4
#> ── Conflicts ───────────────────────────────────────── tidymodels_conflicts() ──
#> x purrr::discard() masks scales::discard()
#> x dplyr::filter()  masks stats::filter()
#> x dplyr::lag()     masks stats::lag()
#> x recipes::step()  masks stats::step()
library(splines)

data(ames)

ames$Sale_Price <- log10(ames$Sale_Price)

set.seed(3)
split <- initial_split(ames)
ames_tr <- training(split)
ames_te <- testing(split)
```

``` r

deg_free <- 10

bs_obj <- bs(ames_tr$Longitude, df = deg_free)
knots <- attributes(bs_obj)$knots
b_knots <- attributes(bs_obj)$Boundary.knots

te_man <- bs(ames_te$Longitude, knots = knots, df = deg_free, Boundary.knots = b_knots)
te_obj <- predict(bs_obj, ames_te$Longitude)
all.equal(te_man, te_obj)
#> [1] TRUE
```

``` r

num_knots <- deg_free - attributes(bs_obj)$degree + 1
knots_manual <- quantile(ames_tr$Longitude, probs = (1:(num_knots-1))/num_knots)
all.equal(knots_manual, knots)
#> [1] TRUE
```

<sup>Created on 2021-02-01 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>
```