@AshesITR 2021-02-02
I'm not sure what you mean.
The point was that directly calling `bs` also evaluates the basis functions, doing a massive useless computation. The PR code omits this part from the `bs` code during prep.

Surely the code in this PR could be refactored a bit to make it more readable, but less similar to the original `bs` code, but directly calling `bs` during prep is the main reason for the poor performance on master.