@topepo 2021-02-02
During `prep()`, we can call `quantile()` and `range()` to get the two quantities that we need. That's what the last code block shows (a point I believe that you had previously mentioned). `bs()` and `ns()` would only get called in `bake()`. Do you agree? 

The code in the PR is almost directly copied from the `splines` package and would violate copywrite. We have to make sure that we do not infringe on other people's work or violate the license. I don't see the need for the new function and don't want to make the R Core team a copyright holder for this part of `recipes` if it can be easily avoided.