@AshesITR 2021-02-02
Unfortunately, if we want to keep all (including edge-case) behaviour, we'll need to duplicate the logic.

e.g. if you specify `Boundary.knots` such that there are some observations out of range, your reprex starts to fail:

``` r
library(tidymodels)
#> -- Attaching packages -------------------------------------- tidymodels 0.1.2 --
#> v broom     0.7.3      v recipes   0.1.15
#> v dials     0.0.9      v rsample   0.0.8 
#> v dplyr     1.0.3      v tibble    3.0.5 
#> v ggplot2   3.3.3      v tidyr     1.1.2 
#> v infer     0.5.4      v tune      0.1.2 
#> v modeldata 0.1.0      v workflows 0.2.1 
#> v parsnip   0.1.5      v yardstick 0.0.7 
#> v purrr     0.3.4
#> -- Conflicts ----------------------------------------- tidymodels_conflicts() --
#> x purrr::discard() masks scales::discard()
#> x dplyr::filter()  masks stats::filter()
#> x dplyr::lag()     masks stats::lag()
#> x recipes::step()  masks stats::step()
library(splines)

data(ames)

ames$Sale_Price <- log10(ames$Sale_Price)

set.seed(3)
split <- initial_split(ames)
ames_tr <- training(split)
ames_te <- testing(split)

deg_free <- 10

bs_obj <- bs(ames_tr$Longitude, df = deg_free, Boundary.knots = c(-93.6, -93.5))
#> Warning in bs(ames_tr$Longitude, df = deg_free, Boundary.knots = c(-93.6, : some
#> 'x' values beyond boundary knots may cause ill-conditioned bases
knots <- attributes(bs_obj)$knots
b_knots <- attributes(bs_obj)$Boundary.knots

te_man <- bs(ames_te$Longitude, knots = knots, df = deg_free, Boundary.knots = b_knots)
#> Warning in bs(ames_te$Longitude, knots = knots, df = deg_free, Boundary.knots =
#> b_knots): some 'x' values beyond boundary knots may cause ill-conditioned bases
te_obj <- predict(bs_obj, ames_te$Longitude)
#> Warning in bs(x = c(-93.6193873, -93.61732, -93.633826, -93.632852,
#> -93.635951, : some 'x' values beyond boundary knots may cause ill-conditioned
#> bases
all.equal(te_man, te_obj)
#> [1] TRUE

num_knots <- deg_free - attributes(bs_obj)$degree + 1
knots_manual <- quantile(ames_tr$Longitude, probs = (1:(num_knots-1))/num_knots)
all.equal(knots_manual, knots)
#> [1] "Mean relative difference: 0.0005032971"
```

<sup>Created on 2021-02-02 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>


Are we okay with breaking some edge cases?

Regardless, I can definitely rewrite the code to reproduce the same logic assuming a non-error situation if that helps with the copyright problem?