@AshesITR 2021-02-02
If we allow for breaking changes and don't want to do the warning / stopping error checks from splines and with the behaviour for ill-conditioned Boundary.knots, we could get away with something like

```
degree <- args$degree %||% switch(type, bs = 3L, ns = 1L)
df <- args$df
intercept <- args$intercept %||% FALSE
boundary <- args$Boundary.knots %||% range(x)

if (!is.null(args$df) && is.null(args$knots)) {
  num_knots <- args$df - degree - intercept
  knots <- quantile(x, (1:(num_knots-1))/num_knots, na.rm = TRUE)
} else {
  knots <- numeric()
}
...
```

would that be okay?